refactor: merge processing.py and processing2.py
